requirements.txt
Python 

PyTorch

NumPy

Librosa

scikit-learn

Matplotlib

pandas